recycleview fix

+++ app/src/main/java/com/example/kata/edrive/network/NetworkItem.java
@@ -4,7 +4,7 @@ package com.example.kata.edrive.network;
  * Created by Kata on 2016. 11. 11..
  */
 public class NetworkItem {
-    public double lat;
-    public double lon;
+    public float lat;
+    public float lon;
     public String city;
 }
